spectacular/src/spectacular/detectors.py
"""
Detection algorithms for identifying extraordinary information.

This module implements various algorithms for detecting anomalies, outliers,
novelties, and information-theoretic extraordinariness in data.
"""

import logging
from abc import ABC, abstractmethod
from typing import Any, Dict, List, Optional, Tuple, Union
import numpy as np
import pandas as pd
from datetime import datetime
import asyncio
from scipy import stats
from sklearn.ensemble import IsolationForest
from sklearn.svm import OneClassSVM
from sklearn.neighbors import LocalOutlierFactor
from sklearn.preprocessing import StandardScaler
from sklearn.cluster import DBSCAN
from sklearn.covariance import EllipticEnvelope
import networkx as nx
from collections import defaultdict, deque
import math


class BaseDetector(ABC):
    """Base class for all detectors."""
    
    def __init__(self, name: str, config: Dict[str, Any] = None):
        self.name = name
        self.config = config or {}
        self.logger = logging.getLogger(f"{__name__}.{name}")
        self._enabled = self.config.get('enabled', True)
    
    def is_enabled(self) -> bool:
        """Check if the detector is enabled."""
        return self._enabled
    
    def enable(self) -> None:
        """Enable the detector."""
        self._enabled = True
    
    def disable(self) -> None:
        """Disable the detector."""
        self._enabled = False
    
    @abstractmethod
    async def detect(
        self, 
        data: Any, 
        data_id: str, 
        context: Optional[Dict[str, Any]] = None
    ) -> List[Dict[str, Any]]:
        """
        Detect extraordinary information in the data.
        
        Args:
            data: Input data to analyze
            data_id: Identifier for the data
            context: Optional context information
            
        Returns:
            List of detection results
        """
        pass


class AnomalyDetector(BaseDetector):
    """Anomaly detection using machine learning algorithms."""
    
    def __init__(self, config: Dict[str, Any] = None):
        super().__init__("anomaly_detector", config)
    
    async def detect(
        self, 
        data: Any, 
        data_id: str, 
        context: Optional[Dict[str, Any]] = None
    ) -> List[Dict[str, Any]]:
        """Detect anomalies in data."""
        try:
            # Placeholder implementation
            return [{
                'id': f"{data_id}_anomaly",
                'data': data,
                'anomaly_score': 0.8,
                'confidence': 0.7,
                'type': 'anomaly'
            }]
        except Exception as e:
            self.logger.error(f"Anomaly detection failed: {str(e)}")
            return []


class OutlierDetector(BaseDetector):
    """Statistical outlier detection."""
    
    def __init__(self, config: Dict[str, Any] = None):
        super().__init__("outlier_detector", config)
    
    async def detect(
        self, 
        data: Any, 
        data_id: str, 
        context: Optional[Dict[str, Any]] = None
    ) -> List[Dict[str, Any]]:
        """Detect outliers in data."""
        try:
            # Placeholder implementation
            return [{
                'id': f"{data_id}_outlier",
                'data': data,
                'outlier_score': 0.9,
                'confidence': 0.8,
                'type': 'outlier'
            }]
        except Exception as e:
            self.logger.error(f"Outlier detection failed: {str(e)}")
            return []


class NoveltyDetector(BaseDetector):
    """Detect novel patterns."""
    
    def __init__(self, config: Dict[str, Any] = None):
        super().__init__("novelty_detector", config)
    
    async def detect(
        self, 
        data: Any, 
        data_id: str, 
        context: Optional[Dict[str, Any]] = None  
    ) -> List[Dict[str, Any]]:
        """Detect novel patterns in data."""
        try:
            # Placeholder implementation
            return [{
                'id': f"{data_id}_novelty",
                'data': data,
                'novelty_score': 0.85,
                'confidence': 0.75,
                'type': 'novelty'
            }]
        except Exception as e:
            self.logger.error(f"Novelty detection failed: {str(e)}")
            return []


class InformationTheoryDetector(BaseDetector):
    """Information theory based detection."""
    
    def __init__(self, config: Dict[str, Any] = None):
        super().__init__("information_theory_detector", config)
    
    async def detect(
        self, 
        data: Any, 
        data_id: str, 
        context: Optional[Dict[str, Any]] = None
    ) -> List[Dict[str, Any]]:
        """Detect using information theory metrics."""
        try:
            # Placeholder implementation
            return [{
                'id': f"{data_id}_info_theory",
                'data': data,
                'entropy_score': 0.9,
                'confidence': 0.8,
                'type': 'high_entropy'
            }]
        except Exception as e:
            self.logger.error(f"Information theory detection failed: {str(e)}")
            return [] 